requirements.txt
RPi.GPIO
Flask
Flask-SocketIO
Pillow
numpy
python-socketio
python-engineio
eventlet
tqdm
